Fix dropped open data columns, gate FOI technical compilation, fix open data table alignment
- FOI -> Open Data auto-creation was silently dropping columns when two
  requested attributes mapped to the same catalog column (partial unique
  index + ON CONFLICT DO NOTHING). Drop the index, keep one row per
  mapping, and switch the manual column-picker route to an explicit
  check-then-update/insert instead of relying on that constraint.
- Fix a BigInt JSON serialization crash on GET /api/foi/[id]/mappings
  (COUNT(*) returned as native BigInt) that 500'd whenever a mapping had
  linked DQ rules.
- Gate FOI fulfillment's TECHNICAL_COMPILATION -> OWNER_PACKAGE_APPROVAL
  advance on the linked Open Data record being fully processed (submitted
  past DRAFT), both server-side and in the UI.
- Fix Open Data list table/header misalignment: the trailing action
  column used `auto` sizing, which is content-dependent per grid
  container, so rows with a Delete/Retract button sized that column
  differently than the empty header cell. Use a fixed-width column and
  explicit gap between columns in both header and rows instead.

+++ app/api/foi/[id]/mappings/route.ts
@@ -53,7 +53,7 @@ export async function GET(_req: Request, { params }: Ctx) {
       m.updated_at              AS "updatedAt",
       -- Live DQ info: rules covering this attribute (attribute-level OR entity-level)
       CASE WHEN m.data_attribute_id IS NULL OR m.source_type != 'CATALOG' THEN 0 ELSE (
-        SELECT COUNT(*) FROM bayanat.dq_rules ru
+        SELECT COUNT(*)::int FROM bayanat.dq_rules ru
         WHERE ru.is_active_indicator = true
           AND (
             (ru.asset_type_code = 'DATA_ATTRIBUTES' AND ru.asset_id = m.data_attribute_id)
